+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com)
 and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [Unreleased] - 
+
+### Added
+
+- Added support for inline styles in Html reader (borders, alignment, width, height)
+
 ## [1.6.0] - 2019-01-02
 
 ### Added